cpu_input_boost: Remove frame boosting entirely [1/2]
At this point, it's begining to seem as through frame boost has
outlived it's usefulness. Currently, it's only being used to raise
the CPU's minimum frequency to 748MHz while rendering.

As of the following commits:
5aea531 ("sched/tune: SchedTune Assist: Increase default top-app boost")
0a99705 ("Revert "cpu_input_boost: Implement scalable Dynamic SchedTune Boosting"")

It's no longer necessary to maintain a seperate boost level for
rendering even in the context of SchedTune. Further more, I think
it would be more viable to bump the minimum frequency to 748MHz
globally and simply allow for scaling down to hardware minimum when
the display is off. This should preserve all of the benefits of
frame boost and then some, while mitigating any negative impact
such a frequency bump may have on idle drain. By taking this
alternative approach, we can substantially reduce the complexity of
this driver while eliminating the constant bouncing between 576 and
748MHz associated with the current frame boost implementation,
which could potentially improve efficiency as well as performance.

This reverts the following commits:
4369c07 ("cpu_input_boost: Allow configuration of each cluster's frame-boost freq")
8f9083c ("cpu_input_boost: Introduce should_kick_frame_boost function")

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>